Client code refactor: Separate types from api and move auth methods into user
