+++ packages/client/api/cardProgressApi.ts
@@ -1,19 +1,11 @@
 import { $axios } from '@/api/baseApi'
-import type { Card } from '@/api'
-import { normalizeCard, type NormalizedCard } from '@/utils'
-
-export interface CardProgress {
-  id: string
-  userId: string
-  cardId: string
-  repetitionCount: number
-  easinessFactor: number
-  intervalDays: number
-  nextReviewAt: number
-  lastReviewedAt: Date
-  card: Card
-}
-export type NormalizedCardProgress = Omit<CardProgress, 'card'> & { card: NormalizedCard }
+import { normalizeCard } from '@/utils'
+import {
+  CardProgress,
+  NormalizedCardProgress,
+  UpdateCardProgressRequest,
+  UpdateCardProgressResponse,
+} from '@/types'
 
 export const getCardsToReview = async (deckId?: string): Promise<NormalizedCardProgress[]> => {
   const res = await $axios.get('/srs/cards', {
@@ -21,9 +13,6 @@ export const getCardsToReview = async (deckId?: string): Promise<NormalizedCardP
   })
   return (res.data as CardProgress[]).map(cp => ({ ...cp, card: normalizeCard(cp.card) }))
 }
-
-type UpdateCardProgressRequest = { grade: number }
-type UpdateCardProgressResponse = Omit<CardProgress, 'card'>
 
 export const updateCardProgress = async (
   cardId: string,
