+++ packages/client/api/cardsApi.ts
@@ -1,33 +1,12 @@
 import { $axios } from '@/api/baseApi'
 import { normalizeCard, type NormalizedCard } from '@/utils'
-
-export interface Card {
-  id: string
-  deckId: string
-  sentence: string
-  createdByUserId: string
-  targetWords: TargetWord[]
-}
-
-export interface TargetWord {
-  id: string
-  word: {
-    id: string
-    text: string
-  }
-}
-
-export interface UserProvidedDefinition {
-  sourceEntryId: string
-  word: string
-  text: string
-  partOfSpeech: string
-  syllabifiedWord: string
-  pronunciationAudioUrl?: string
-  offensive?: boolean
-  labels?: string[]
-  stems?: string[]
-}
+import {
+  Card,
+  CreateCardRequest,
+  CreateCardResponse,
+  EditCardRequest,
+  EditCardResponse,
+} from '@/types'
 
 export const getCardsByDeckId = async (
   deckId: string,
@@ -50,29 +29,12 @@ export const deleteCard = async (cardId: string) => {
   return await $axios.delete(`/cards/${cardId}`)
 }
 
-export interface CreateCardRequest {
-  sentence: string
-  targetWords: string[]
-  definitions?: UserProvidedDefinition[]
-}
-interface CreateCardResponse {
-  card: Card
-  notFoundWords: string[]
-  inserted: boolean
-}
-
 export const createCard = async (
   deckId: string,
   data: CreateCardRequest
 ): Promise<CreateCardResponse> => {
   const res = await $axios.post(`/cards/deck/${deckId}`, data)
   return res.data
-}
-
-export type EditCardRequest = Partial<CreateCardRequest>
-interface EditCardResponse {
-  card: Card
-  notFoundWords: string[]
 }
 
 export const editCard = async (
