+++ packages/client/api/definitionApi.ts
@@ -1,32 +1,5 @@
 import { $axios } from '@/api/baseApi'
-
-export interface Definition {
-  id: string
-  sourceEntryId: string
-  wordId: string
-  audio?: string
-  text: string
-  partOfSpeech:
-    | 'noun'
-    | 'pronoun'
-    | 'verb'
-    | 'adjective'
-    | 'adverb'
-    | 'phrasal verb'
-    | 'idiom'
-    | 'preposition'
-    | 'conjunction'
-    | 'interjection'
-  labels: string[]
-  syllabifiedWord: string
-  offensive: boolean
-  source: 'dictionary' | 'user'
-  sourceName?: string
-  approved: boolean
-  createdByUserId: string
-}
-
-type GetDefinitionsForCardResponse = Definition[]
+import { GetDefinitionsForCardResponse } from '@/types'
 
 export const getDefinitionsForCard = async (
   cardId: string,
