+++ packages/client/api/index.ts
@@ -1,4 +1,3 @@
-export * from './authApi'
 export * from './userApi'
 export * from './deckApi'
 export * from './cardsApi'
